backend/requirements.txt
# Voice-Flow Backend Requirements
# For Parakeet TDT 0.6B v2 STT Model

# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiohttp==3.11.11

# Database
sqlalchemy==2.0.36
pydantic==2.10.5
pydantic-settings==2.7.0

# Environment
python-dotenv==1.0.1

# Firebase Authentication
firebase-admin==6.5.0

# PDF/CSV export
reportlab==4.2.5

# Audio processing
pydub==0.25.1
librosa==0.11.0
soundfile==0.13.1

# CRITICAL: NumPy must be < 2.0 for NeMo compatibility
# NumPy 2.0+ removes np.sctypes which NeMo uses
numpy>=1.26.0,<2.0

# PyTorch with CUDA 12.1 support
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchaudio==2.5.1+cu121

# NeMo ASR (for Parakeet model)
nemo-toolkit==2.1.0
omegaconf==2.3.0
hydra-core==1.3.2
sentencepiece==0.2.1
einops==0.8.1
lightning==2.6.0

# NeMo dependencies
pytorch-lightning==2.6.0
datasets
pandas
matplotlib
ipython
pyannote.core
pyannote.metrics
jiwer
nltk
inflect
peft
kaldialign
lhotse
webdataset
fiddle
cloudpickle
editdistance

# Testing
pytest==8.3.4
httpx==0.28.1

# PyInstaller for exe building
pyinstaller
